refactor: extract ProcessInstanceMapper as an instantiated per-entity mapper
Moves all process-instance REST mapping (create, cancel, suspend, resume,
migrate, modify, and their batch-operation variants) out of the monolithic
RequestMapper into a dedicated ProcessInstanceMapper, following the
GroupMapper/TenantMapper/AuthorizationMapper pattern already established in
gateways/gateway-mapping-http/.../mapper/: an instantiated class held by
ProcessInstanceController, rather than a bag of static methods.
