+++ gateways/gateway-mapping-http/src/test/java/io/camunda/gateway/mapping/http/rest/RequestMapperTest.java
@@ -10,7 +10,6 @@ package io.camunda.gateway.mapping.http.rest;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.gateway.mapping.http.RequestMapper;
-import io.camunda.gateway.protocol.model.AdvancedStringFilter;
 import io.camunda.gateway.protocol.model.DeleteResourceRequest;
 import io.camunda.gateway.protocol.model.JobActivationRequest;
 import io.camunda.gateway.protocol.model.JobChangeset;
@@ -18,165 +17,14 @@ import io.camunda.gateway.protocol.model.JobCompletionRequest;
 import io.camunda.gateway.protocol.model.JobErrorRequest;
 import io.camunda.gateway.protocol.model.JobFailRequest;
 import io.camunda.gateway.protocol.model.JobUpdateRequest;
-import io.camunda.gateway.protocol.model.MigrateProcessInstanceMappingInstruction;
-import io.camunda.gateway.protocol.model.ProcessInstanceFilter;
-import io.camunda.gateway.protocol.model.ProcessInstanceMigrationBatchOperationPlan;
-import io.camunda.gateway.protocol.model.ProcessInstanceMigrationBatchOperationRequest;
-import io.camunda.gateway.protocol.model.ProcessInstanceModificationBatchOperationRequest;
-import io.camunda.gateway.protocol.model.ProcessInstanceModificationMoveBatchOperationInstruction;
 import io.camunda.gateway.protocol.model.TenantFilterEnum;
 import io.camunda.service.JobServices.UpdateJobChangeset;
-import io.camunda.service.ProcessInstanceServices.ProcessInstanceMigrateBatchOperationRequest;
-import io.camunda.service.ProcessInstanceServices.ProcessInstanceModifyBatchOperationRequest;
 import io.camunda.zeebe.protocol.record.value.TenantFilter;
-import io.camunda.zeebe.util.Either;
 import java.util.List;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.ProblemDetail;
 
 class RequestMapperTest {
-
-  @Test
-  void shouldMapToProcessInstanceMigrationBatchOperationRequest() {
-    // given
-    final var mappingInstruction =
-        MigrateProcessInstanceMappingInstruction.Builder.create()
-            .sourceElementId("source1")
-            .targetElementId("target1")
-            .build();
-    final var migrationPlan =
-        ProcessInstanceMigrationBatchOperationPlan.Builder.create()
-            .targetProcessDefinitionKey("123")
-            .mappingInstructions(List.of(mappingInstruction))
-            .build();
-    final var filter =
-        ProcessInstanceFilter.Builder.create()
-            .processDefinitionId(AdvancedStringFilter.Builder.create().$like("process").build())
-            .build();
-    final var batchOperationInstruction =
-        ProcessInstanceMigrationBatchOperationRequest.Builder.create()
-            .filter(filter)
-            .migrationPlan(migrationPlan)
-            .build();
-
-    // when
-    final Either<ProblemDetail, ProcessInstanceMigrateBatchOperationRequest> result =
-        RequestMapper.toProcessInstanceMigrationBatchOperationRequest(batchOperationInstruction);
-
-    // then
-    assertThat(result.isRight()).isTrue();
-    final var request = result.get();
-    assertThat(request.targetProcessDefinitionKey()).isEqualTo(123L);
-    assertThat(request.mappingInstructions())
-        .hasSize(1)
-        .first()
-        .satisfies(
-            instruction -> {
-              assertThat(instruction.getSourceElementId()).isEqualTo("source1");
-              assertThat(instruction.getTargetElementId()).isEqualTo("target1");
-            });
-  }
-
-  @Test
-  void shouldReturnProblemDetailForInvalidInput() {
-    // given
-    final var mappingInstruction =
-        MigrateProcessInstanceMappingInstruction.Builder.create()
-            .sourceElementId("")
-            .targetElementId("")
-            .build();
-    final var migrationPlan =
-        ProcessInstanceMigrationBatchOperationPlan.Builder.create()
-            .targetProcessDefinitionKey("123")
-            .mappingInstructions(List.of(mappingInstruction))
-            .build();
-    final var filter = ProcessInstanceFilter.Builder.create().build();
-    final var batchOperationRequest =
-        ProcessInstanceMigrationBatchOperationRequest.Builder.create()
-            .filter(filter)
-            .migrationPlan(migrationPlan)
-            .build();
-
-    // when
-    final Either<ProblemDetail, ProcessInstanceMigrateBatchOperationRequest> result =
-        RequestMapper.toProcessInstanceMigrationBatchOperationRequest(batchOperationRequest);
-
-    // then
-    assertThat(result.isLeft()).isTrue();
-    final var problemDetail = result.getLeft();
-    assertThat(problemDetail.getStatus()).isEqualTo(400); // Bad Request
-    assertThat(problemDetail.getDetail()).contains("are required");
-  }
-
-  @Test
-  void shouldMapProcessInstanceModifyBatchOperationRequest() {
-    // given
-    final var moveInstruction =
-        ProcessInstanceModificationMoveBatchOperationInstruction.Builder.create()
-            .sourceElementId("source1")
-            .targetElementId("target1")
-            .build();
-    final var filter =
-        ProcessInstanceFilter.Builder.create()
-            .processDefinitionId(AdvancedStringFilter.Builder.create().$like("process").build())
-            .build();
-    final var modificationRequest =
-        ProcessInstanceModificationBatchOperationRequest.Builder.create()
-            .filter(filter)
-            .moveInstructions(List.of(moveInstruction))
-            .build();
-
-    // when
-    final Either<ProblemDetail, ProcessInstanceModifyBatchOperationRequest> result =
-        RequestMapper.toProcessInstanceModifyBatchOperationRequest(modificationRequest);
-
-    // then
-    assertThat(result.isRight()).isTrue();
-    final var request = result.get();
-    assertThat(request.moveInstructions())
-        .hasSize(1)
-        .first()
-        .satisfies(
-            instruction -> {
-              assertThat(instruction.getSourceElementId()).isEqualTo("source1");
-              assertThat(instruction.getTargetElementId()).isEqualTo("target1");
-              assertThat(instruction.getAncestorScopeKey()).isEqualTo(-1L);
-              assertThat(instruction.isInferAncestorScopeFromSourceHierarchy()).isTrue();
-              assertThat(instruction.isUseSourceParentKeyAsAncestorScopeKey()).isFalse();
-              assertThat(instruction.getVariableInstructions()).isEmpty();
-            });
-  }
-
-  @Test
-  void shouldNotMapProcessInstanceModifyBatchOperationRequestWhenInvalid() {
-    // given
-    // Use empty targetElementId to trigger "No targetElementId provided." validation
-    final var moveInstruction =
-        ProcessInstanceModificationMoveBatchOperationInstruction.Builder.create()
-            .sourceElementId("source1")
-            .targetElementId("")
-            .build();
-    final var filter =
-        ProcessInstanceFilter.Builder.create()
-            .processDefinitionId(AdvancedStringFilter.Builder.create().$like("process").build())
-            .build();
-    final var modificationRequest =
-        ProcessInstanceModificationBatchOperationRequest.Builder.create()
-            .filter(filter)
-            .moveInstructions(List.of(moveInstruction))
-            .build();
-
-    // when
-    final Either<ProblemDetail, ProcessInstanceModifyBatchOperationRequest> result =
-        RequestMapper.toProcessInstanceModifyBatchOperationRequest(modificationRequest);
-
-    // then
-    assertThat(result.isLeft()).isTrue();
-    final var problemDetail = result.getLeft();
-    assertThat(problemDetail.getStatus()).isEqualTo(400);
-    assertThat(problemDetail.getDetail()).isEqualTo("No targetElementId provided.");
-  }
 
   @Nested
   class ResourceDeletionRequestMappingTest {
